Added Mainmenu with player name input and play button.
made a map where the player cna input their name, the name gets saved in a game instance and whenever a new game mode starts it takes that name from the instance. WHen we print the name of the data file we take that name from the game mode. The reason I dont get the infor directly from the game instacne is because I was lazy and did not make it a c++ class.